Played with classNameBindings, got it to toggle new-vehicle component
CamelCase in component, with-dashes in scss file! That screwed me up for a
while. Alas, I don't think this is actually how I want users to create new
vehicles. Too cluttered a view this way. So I'm going to refactor so that
it's like update - a new route will have the job of rendering the new component