Create successful createWork form and action
At first my work model was missing the vehicleId property, and that needed
to be written in the model or Ember wouldn't know what I was talking about.
And for a long time I was getting errors because I wrote the property as
vehicle_id in the Ember model (since that's what the database expects).
Unbeknownst to be, Ember wants you to use camelCase, and does its magic to convert
vehicleId into vehicle_id for the database. Lesson learned!