Finally got the Ember relationship between vehicle and works set up
So having the work model have an attribute `  vehicle: DS.belongsTo('vehicle')`
instead of `vehicleId: DS.attr('string')` was key to being able to access
vehicle.works in Ember just like you can in the Rails console.

Of course, vehicle model has a corresponding `  works: DS.hasMany('work')` line

The issue I was having before, and the reason I wasn't using the belongsTo on
my model, is that when I was creating works, I was trying to pass a vehicleId
along with the work. I got it working one way, but not with the belongsTo relationship

With the belongsTo relationship, you're not supposed to pass the id, you're supposed
to pass the WHOLE Ember object! So even though the database is expecting a vehicle_id
field on Work, in Ember you don't create a vehicleId property, you create a property
simply called `vehicle`, which gets passed the entire object. It then works its magic
to get the id of that object and give it to the database