Bug fix render vehicle-works in /vehicle route, add deleteWork
I was having a bug where the vehicle-works component wasn't loading on the
/vehicle route until the /works route was hit and I went back to the /vehicle
route. But in doing so it would un-render the vehicle make model and year.

The issue was that my /vehicle route model hook was never being called! Ember
was seeing that I had a model called vehicle and was assuming RESTful convention
and going and getting that model. That's why the RSVP hash stuff didn't work either.

Anyway, problem was solved simply by commenting out (soon to be deleted) model hook.
In the vehicle route, {{model.works}} is an array of the works belonging to that model!
Since I want extra funcitonality like deleting and editing those works, having the component
vehicle-works is still a good way to go.

Created a deleteWork button and action and it's copacetic