Successfully created an update-work route and update-form component
Vehicle view now has a link to an update work route that renders an
update work form component.

The only weird thing is that ember is doing another implicit model hook
for my update work route (maybe since its path is /work/:work_id)
and its overwriting the explicit one I wrote.

The only functionality that's lacking as the result of this weirdness
is that I can't do a transitionTo after updating the work. But I've
spent way too much time on this, and since there's a working link-to
button that does exactly what the transitionTo would do, I'm going to move
on