Improve layout of vehicle route by changing bootstrap
my-application was rendering {{outlet}} inside of a bootstrap div
that had an offset-column property (this was set up from the GA ember
template). This was forcing that styling into all other pages, so I deleted that
and in its place I added more custom boostrap columns to the vehicle view.
(The vehicles view was unnaffected by the change since everything is centered
on the page anyway).

I added some headers for the vehicle works, changed some header sizes around,
added borders and margins to works so that they're easy to read and essentially
resemble a table.

Since they resemble a table I may experiment with the HTML <table> tag, though
there's nothing inherently buggy or bad about the styling as is.