Create helper for putting commas in mileage readings in vehicle route
Making the helper work for vehicle.mileage was easy enough

Ran into trouble when using the helper for works.mileage

Since works are displayed as a component in the vehicle route, I was
running into an async issue, where even though work.mileage was rendering
on the page, the helper was getting to it before the request for the work was made,
and was receiving undefined. Since you can't add commas to undefined, crazy things
like an infinite rendering happened.

Fixed this by writing an explicit model hook for the vhicle route with an RSVP
hash that makes a GET for the vehicle and for All works

TODO: change the RSVP model hook to make a GET to only the works that belong to the vehicle
This is done with query params. Check out issue
https://git.generalassemb.ly/ga-wdi-boston/capstone-project/issues/1125#issuecomment-87358

The Rails index action will see something like GET/works?vehicle=4

and I need some code like if params[:vehicle] === 4, do
    find works where vehicle_id is 4
else
    find all current users works

For an app of my size I'm ok with getting all works for now, but as soon as app grows
this could be a problem.