Fix bug where on update, transitionTo vehicle route didn't load vehicle
I'm almost positive this was working in previous commits, but maybe when
I wrote an explicit model hook for the vehicle route which returned an RSVP hash
something broke. Even though I was passing the same exact thing to the update route,
I was now calling it model.vehicle instead of just model.

I called model.vehicle simply vehicle in the update route. Yet for some reason
passing vehicle.id to the transitionTo fixed the bug, where before passing vehicle
to transitionTo was all that was needed. Maybe it was always a race condition, but now that
the vehicle model hook gets all works as well, it takes longer, and the transitionTo happens before
it's returned. Either way, it seems fixed now.

The bug was that the vehicle wasn't loaded so anything pertaining to vehicle.<property>
including vehicle.works just wasn't rendering