Create splash page with sign in and up forms
Created a splash page on index with sign in/up forms
that toggle visibility when buttons are clicked

Added a car picture background. In order to make the body background
a picture here but not on other routes, I used 'activate' and
'deactivate' properties on the index route. For some reason these
didn't work on the application route (probably because application is
the root) and they don't seem to work on components (or at least you
can't target the body on components. Had no luck targeting other elements
by ID either though).

Had to define signin and signup methods for the index route (copied
them from their respective routes) and included auth and a model hook
for index to get them working. The form fields for email and password had
IDs, and ember didn't like the fact that I was trying to display two
components on one page with the same ids, so since those ids weren't being
used for anything, I changed them to class=email and class=password.