Move sign up form into sign in form component
Sign-in-form component now has both forms, and actions for both
forms that it passes up to index. This seemed to be the best way
to have both forms overlap (take up the same space on the screen)
the only downside is now the name of the component isn't totally
accurate. I could generate a new component, paste all the code into it,
then clean up the unused components, but that's for when I'm cleaning
up my code.

Got the styling pretty much how Kaitlyn suggested it. Bootstrap buttons
leave annoying blue border when clicked, and forms aren't the sleek,
single lines, but it'll get there!