Add my-application component (navbar) to routes
Since I don't want the navbar on the splash page I'm not rendering
my-application component in the application route or index route.
Therefore it needs to be added to other routes.

Added it to 'vehicles' route and 'vehicle' route.

Changed some transitionTos to hit index or vehicles instead of sign-up
route (which I'm no longer using)

Added some styling to navbar

TODO: when app is more or less finished, clean up unused routes