Add price to works model and view
Added price to works model and as a new column in the vehicle-works
component.
Added price to new-work form and update-work form.
Created helper to format price with a `$` and 2 decimal places

Also moved update and delete work buttons to the right of the components
instead of as a new row.