Bugfix: format-mileage now accepts null values
Previously, format-milage would break the page if a user created a work
without a value for mileage. I got around this by making mileage a
required property, even though ideally I would let users log works with-
out mileage (because maybe only the date is important).

The problem was, when `mileage` is being sent (on create or on update)
without a value, its value is set to `undefined`.
However, when `mileage` is being read by the `vehicle-works` component,
it’s value is `null` (if it was `undefined` when created/updated)

If the helper wasn’t checking that `mileage !== undefined`, it would
throw the error `cannot read property .toString of undefined`. That was
easily fixed with an `if` statement.

The half of the equation I couldn’t solve before was that if the helper
wasn’t checking that `mileage !== null` it would make an infinite render
loop! I couldn't figure out why the infinite render loop was occurring.
It was only after creating a helper for `price` and seeing
its value as `null` that I wondered why it wasn’t undefined. Now I know!
Funnily enough, it didn’t care if price was `null` (it just wouldn’t
show anything). So there must be something in the format-mileage helper
that royally screws up when given a null value.